订单管理

+++ admin-dao/src/main/java/com/woollen/admin/dao/entry/OrderInfo.java
@@ -74,7 +74,7 @@ public class OrderInfo extends Model<OrderInfo> {
     /**
      * 支付时间
      */
-    private Integer payTime;
+    private Long payTime;
 
     /**
      * 创建时间
